// src/stores/modules/types/purchase.ts
/*
 * @Author: sdpzhong
 * @Date: 2025-03-21 16:04:26
 * @Last Modified by: sdpzhong
 * @Last Modified time: 2025-03-24 14:24:26
 */

export enum SkuType {
  combo = 'COMBO',
  cs = 'CS',
  gsmdata = 'DATA',
}

// 流量套餐类型
export enum TrafficType {
  // 无限
  unlimited = 0,
  // 有限
  limited = 1,
}